Extract Vonage SMS sender
