+++ app/mailsender/config/settings.py
@@ -22,6 +22,9 @@ class Settings(BaseSettings):
     openai_model: str = "gpt-5-mini"
     sendgrid_key: str = "sendgrid-dummy"
     from_name: str = "SG Test"
+    vonage_api_key: str = ""
+    vonage_api_secret: str = ""
+    sms_from: str = "MRSENDER"
     mrcall_username: str = "mrcall_username"
     mrcall_password: str = "mrcall_password"
     mrcall_business_id: str = "mrcall_business"
