+++ app/resources/settings_template.ini
@@ -10,3 +10,6 @@ email_prompt = Sei un assistente che scrive email molto semplici e naturali. Cre
 body = "<p>Hi, this is the default body.</p>"
 database_url = sqlite:///./mailsender.db
 body = Ciao {first}, <br><br>questo è un corpo email di esempio.<br><br>Saluti,<br>MrCall <br><br>
+vonage_api_key = your-vonage-key
+vonage_api_secret = your-vonage-secret
+sms_from = MRSENDER
